requirements.txt
alembic==1.16.2
anyio==4.9.0
asyncpg==0.30.0
fastapi==0.115.14
httpx==0.28.1
Jinja2==3.1.6
MarkupSafe==3.0.2
pandas==2.3.0
pgvector==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
sentence-transformers==5.0.0
SQLAlchemy==2.0.41
starlette==0.46.2
torch==2.7.1
transformers==4.53.0
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
psycopg2-binary

